dataset-pipeline/requirements.txt
arxiv
requests
beautifulsoup4
pandas
tqdm
transformers
datasets
ollama
google-generativeai
python-dotenv
